修复了一些问题 新增了职业(下个提交新增更多职业)

+++ src/main/java/net/mcbbs/jerryzrf/minehunt/util/StatisticsBaker.java
@@ -16,7 +16,7 @@ public class StatisticsBaker {
         if (result.getValue() == 0.0d) {
             return "";
         }
-        return result.getKey() + " 对其他生物造成了 " + result.getValue().intValue() / 10 + " 点伤害";
+        return result.getKey() + " 对其他生物造成了 " + result.getValue().doubleValue() / 10.0 + " 点伤害";
     }
     
     public String getDamageTakenMaster() {
@@ -24,7 +24,7 @@ public class StatisticsBaker {
         if (result.getValue() == 0.0d) {
             return "";
         }
-        return result.getKey() + " 共受到了 " + result.getValue().intValue() / 10 + " 点伤害";
+        return result.getKey() + " 共受到了 " + result.getValue().doubleValue() / 10.0 + " 点伤害";
     }
     
     public String getWalkingMaster() {
@@ -32,7 +32,7 @@ public class StatisticsBaker {
         if (result.getValue() == 0.0d) {
             return "";
         }
-        return result.getKey() + " 旅行了 " + result.getValue().intValue() / 100 + " 米";
+        return result.getKey() + " 旅行了 " + result.getValue().doubleValue() / 100.0 + " 米";
     }
